+++ core/CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased] <!-- release-date -->
 
+### Changed
+
+- Remove unused dependency `secrecy`.
+
 ## [0.3.3] - 2024-04-05
 
 ### Added